+++ packages/orq-rc/src/models/operations/createprompt.ts
@@ -888,8 +888,8 @@ export type CreatePromptResponseBody = {
   domainId: string;
   created: string;
   updated: string;
-  createdById?: string | undefined;
-  updatedById?: string | undefined;
+  createdById?: string | null | undefined;
+  updatedById?: string | null | undefined;
   /**
    * The prompt’s name, meant to be displayable in the UI.
    */
@@ -3722,8 +3722,8 @@ export const CreatePromptResponseBody$inboundSchema: z.ZodType<
   domain_id: z.string(),
   created: z.string(),
   updated: z.string(),
-  created_by_id: z.string().optional(),
-  updated_by_id: z.string().optional(),
+  created_by_id: z.nullable(z.string()).optional(),
+  updated_by_id: z.nullable(z.string()).optional(),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   prompt_config: z.lazy(() => CreatePromptPromptsPromptConfig$inboundSchema),
@@ -3747,8 +3747,8 @@ export type CreatePromptResponseBody$Outbound = {
   domain_id: string;
   created: string;
   updated: string;
-  created_by_id?: string | undefined;
-  updated_by_id?: string | undefined;
+  created_by_id?: string | null | undefined;
+  updated_by_id?: string | null | undefined;
   display_name: string;
   description?: string | null | undefined;
   prompt_config: CreatePromptPromptsPromptConfig$Outbound;
@@ -3767,8 +3767,8 @@ export const CreatePromptResponseBody$outboundSchema: z.ZodType<
   domainId: z.string(),
   created: z.string(),
   updated: z.string(),
-  createdById: z.string().optional(),
-  updatedById: z.string().optional(),
+  createdById: z.nullable(z.string()).optional(),
+  updatedById: z.nullable(z.string()).optional(),
   displayName: z.string(),
   description: z.nullable(z.string()).optional(),
   promptConfig: z.lazy(() => CreatePromptPromptsPromptConfig$outboundSchema),