+++ packages/orq-rc/RELEASES.md
@@ -659,3 +659,13 @@ Based on:
 - [typescript v3.2.0-rc.26] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.26 - packages/orq-rc
+
+## 2025-03-04 22:55:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.509.1 (2.539.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.27] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.27 - packages/orq-rc